Add ANY logic for local builds.
For ANY paths, we find the first dev or test image that exists in the
local dir and return that path. This is useful for developers who may
have a dev or test image and more flexible than our previous assumption
that all devs wanted test images.

BUG=chromium:275702
TEST=unittests + 2 updates via AU client one with just a dev image, and one
with just a test image. Both succeeded!

Change-Id: I97897ab10804df4f77e40806706f05a647db8eb6
Reviewed-on: https://chromium-review.googlesource.com/66208
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>