devserver: make artifact names globs by default
What this means is that, by default, we'll be using a glob expression
(i.e. a path that may contain shell-style wildcards) instead of a regex
to identify artifacts. We are still allowing artifacts to be defined by
a regex, as this is useful for some purposes such as disallowing certain
file name patterns; this semantics, however, needs to be specifically
selected via a flag.

Some additional changes included:

* Eliminate enforcing a single file in gsutil_util.GetGSNamesWithWait();
  this is already done in the caller function and there's no need to
  complicate the otherwise generic logic in this module.

* Have DeltaPayloadsArtifact override various name filtering parameters
  internally; letting the user define these may cause inconsistencies
  (especially given the glob vs regex semantics), makes it less robust
  and is not needed in practice.

* Fixes passing of dictionary args (as opposed to list args) to artifact
  objects via artifact factory.

* Fixes a great many gpylint errors/warnings.

BUG=chromium:280220
TEST=Unit tests
TEST=gpylint

Change-Id: I9ea59b7f8962a71f7b387b99b0577932e753c2eb
Reviewed-on: https://chromium-review.googlesource.com/167434
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>
Commit-Queue: Gilad Arnold <garnold@chromium.org>