[autotest] Create a DevServerHTTPError Exception class.
Cherrypy handles it's HTTPErrors by generating html and displaying
the exception message in the browsers view. To get the same exception
message in the logs, first log the actual exception then call
cherrypy.HTTPErrors. This will add a lot of traceability to
HTTPExceptions when we eventually dump the update engine logs in
autoserv.DEBUG.

TEST=Raised an exception and check that it was logged, and displayed
     in the browser view.
BUG=chromium:279299

Change-Id: I20dc817acbb770b6ab44cfea7fa97392a0c20f6b
Reviewed-on: https://chromium-review.googlesource.com/167422
Tested-by: Prashanth Balasubramanian <beeps@chromium.org>
Reviewed-by: Alejandro Deymo <deymo@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Prashanth Balasubramanian <beeps@chromium.org>