devserver: Make symlink creation atomic.
Without this patch, the symlink creation is performed in two steps:
if the symlink already exists, it is first removed and then it
re-created. When two or more threads are creating the same symlink
in this wat, a race condition could make a thread attempt to create
a symlink that already exists because another thread created it
between the two steps. This situation is faced on the referenced bug.

On the other hand, a two-step symlink creation when the symlink
already exists leaves a small window of time where the symlink to
be replaced doesn't exists. This could make another thread trying
to read the symlink (either the previous or the new one) fail.

This patch fixes both problems by creating the symlink atomically.
This is done in a standard way, creating it somewhere else on the
same file system and moving it to its final destination.

BUG=chromium:270434
TEST=common_util unittest the symlink creation.

Change-Id: I38f24adcc5bf2f6d289105697de3981ecd51ad38
Reviewed-on: https://chromium-review.googlesource.com/167971
Reviewed-by: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>