Properly direct download progress diagnostic
All of the "echo" diagnostics in the stateful_update script
are supposed to be redirected to stderr.  One was missed.
This made it difficult to diagnose a recent bug experienced
by image_to_live.  This patch properly redirects the "echo"
progress-diagnostic message to stderr.

BUG=chromium:279621
TEST=image_to_live

Change-Id: Ifad1257f9398a684ed41a96535d7d1059baa75e3
Reviewed-on: https://chromium-review.googlesource.com/168220
Reviewed-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Tested-by: Paul Drews <paul.drews@intel.com>
Commit-Queue: Paul Drews <paul.drews@intel.com>