cbf: restore ability to bundle 'small' images
Passing parameter '-s' to cros_bundle_firmware causes it to try to add
to the image all flashmap sections, even those which are not marked
'required'.

This causes a failure, because the list of sections allowed in the
image is limited to those which are marked required. This change makes
sure that sections which are not required are not attempted to be
added to the final image.

BRANCH=None
BUG=None
TEST=manual
   . reemerged cros-devutils package

   . ran emerge-daisy chromeos-u-boot chromeos-bootimage, observed a
     4M binary images created as expected.

   . ran cros_bundle_firmware manually, passing it '-s' as one of the
     parameters.

     It was failing before this change, and succeeded now. The image
     is created is 700K in size, the reason being that the recovery
     u=boot is marked as required in the flash map. This will be fixed
     separately.

   . attempts to run portage unittests failed due to unrelated tests
     being broken.

Change-Id: I515994290a806ab9a3cb4134d0776463addd8eea
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/169451
Reviewed-by: Simon Glass <sjg@chromium.org>