Allow device trees without revison map
Not all platforms' device trees include board revision map, one of the
examples of such platform is spring.

Just skip adding the revision map to the SPL blob if the map is not in
the device tree. Print an info message instead.

BRANCH=None
BUG=chrome-os-partner:23024
TEST=manual
  . emerge-daisy_spring chromeos-bootimage

Change-Id: Ic133a145c5993718667d4754dee7354ab1569926
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/172164
Reviewed-by: Doug Anderson <dianders@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>