Allow xbuddy to accept any remote urls (including trybot urls).
This changes xbuddy to attempt to lookup a remote board by using the remote
path as is before trying it with the commonly used + -release suffix.

This also makes xbuddy fail faster if someone tries to use a remote url that
doesn't exist since the exception is raised during url resolution rather than
when we are trying to use the artifact.

BUG=chromium:306609
TEST=image_to_live with --image=xbuddy:remote/trybot-test-ap/R32-4777.0.0-b0
+ unit tests.

Change-Id: I352cd93ca2af2b048353dc349b08f734cf8095e7
Reviewed-on: https://chromium-review.googlesource.com/172764
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Chris Sosa <sosa@chromium.org>
Tested-by: Chris Sosa <sosa@chromium.org>