Fix the offset passed into 'google,board-rev-map'
We were accidentally passing in the offset for 'google,board-rev-map'
as the offset from right after the "params" field.  This is a little
awkward, so instead change it to calculate based on the start of the
machine parameters.

BRANCH=pit
BUG=chrome-os-partner:23024
TEST=With other CLs, SPL can now find the proper revision

Change-Id: I105dc2a8c5f2c3b6d673fb83bdb8ea1f36386c65
Signed-off-by: Doug Anderson <dianders@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/173238
(cherry picked from commit 54542e809a16d219623aaf9fb090cfe8dfffab65)
Reviewed-on: https://chromium-review.googlesource.com/173346
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>