Improve wilis robustness
willis spawns multiple threads, and the main thread does not terminate
until all the spawned threads terminate. This causes a problem when an
exception happens on a thread - it never properly terminates, and
willis hangs forever.

This patch adds some code to indicate to the user that an error has
happened. Processing could be improved, the important thing is that
the utility does not hang in case of a problem.

BUG=None
TEST=manual
  . with the recent change modifying the tracking branch property name
    in the project object, all threads spawned by willis throw
    KeyError exceptions, never terminating. With this fix the utility
    reports the error and terminates.

Change-Id: I1093f40d99fb8be12dd39eaa7c603fb3229880cc
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/177285
Reviewed-by: David James <davidjames@chromium.org>