Add 'relative_path' to xbuddy to return relative path of the payload folder
When relative_path is used alone, devserver returns the relative path
(based on the static directory) to the folder where the payloads
are. When it is used with for_update, devserver pre-generates the
payloads and returns the relative path. This is intended for other
scripts to discover the actual location of the payloads, assuming the
static directory is known by the script beforehand.

E.g.
  http://devserver:port/xbuddy/local/peppy/latest/test?relative_path=true
returns
  peppy/R33-5093.0.2013_12_12_1156-a1

BUG=None
TEST=devserver_integration_test.py
TEST=manual testing with a local devserver

Change-Id: I8722d6b5bf446ce05f29bb54ca88ed66a5876582
Reviewed-on: https://chromium-review.googlesource.com/179920
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Yu-Ju Hong <yjhong@chromium.org>
Tested-by: Yu-Ju Hong <yjhong@chromium.org>