bundle_firmware: Update Tegra entries and isolate Exynos a bit
This script seemed to still be using the tegra20.dtsi file, and had
references to tegra20. Currently the only supported tegra platform will
be Nyan which is a Tegra124 platform. Hence update the arch_dts. Also
update the default flashmap name for tegra to nyan.

Also, remove some confusing text about this being a Tegra script.
While it might have started out that way, it is certainly more
general now, and the text was misleading.

Lastly, only some platforms have need of the /iram and /config
nodes. It seems unique to exynos platforms. On other platforms
like Nyan, it simply causes excess warnings. As it was working
without a flashmap, and hence without these nodes already, this
should be fine for Tegra. This does however also block this code path
for coreboot.

BUG=chrome-os-partner:22596
TEST=Verified I can flash nyan platform with u-boot and receive no
warnings.

$ cros_bundle_firmware --board=nyan --write=usb --flash=spi:1 --servo=none \
   --dt=/build/nyan/firmware/dtb/tegra124-venice2.dtb

No warnings displayed, and flash was successful.

Change-Id: I204af8d0b625d597ad8d8637fbfb930617594933
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/181642
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Katie Roberts-Hoffman <katierh@chromium.org>