Handle channel lookup correctly when board name includes underscore(s)
For board names with underscore(s) in it (e.g. daisy_spring), the
canary builds are archived in:
  gs://chromeos-releases/dev-channel/daisy-spring/5000.0.0/

However, due to historical reasons, the underscores in the board name
are replaced with hyphens in the corresponding channels of the board:
  gs://chromeos-image-archive/daisy_spring-release/

This CL converts the board name to use hyphens before performing the
channel lookup.

BUG=chromium:336711
TEST=`cros flash device xbuddy://remote/daisy_spring/latest-canary`

Change-Id: I26a3fa392d19459d78ed4c4de73c5af08b7a4e84
Reviewed-on: https://chromium-review.googlesource.com/183440
Tested-by: Yu-Ju Hong <yjhong@chromium.org>
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Yu-Ju Hong <yjhong@chromium.org>