xbuddy: remove DEV from GS_ALIAS
This would fix the alias-to-filename mapping because there are no
correspoinding entries in GS_FILE_NAMES and ARTIFACTS for the dev
image. The alternative is to add the missing entries, but we cannot
produce dev image anyway because it is not included in image.zip.

BUG=chromium:342196
TEST=None

Change-Id: Ic53484f291f141f224997f4cd88275ff66900321
Reviewed-on: https://chromium-review.googlesource.com/185668
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Queue: Yu-Ju Hong <yjhong@chromium.org>
Tested-by: Yu-Ju Hong <yjhong@chromium.org>