cros_generate_update_payload: Succeed if retry works.
Currently, we return a failure result even if the rety cleanup works in
cros_generate_update_payload. This was done to make it easier to find
retries in build results.

This changes us to succeeding if the retry works, as long as it works
eventually.

BUG=chromium:353558
TEST=Local testing + tryjob.

Change-Id: I4dd526a03bb31156cf175e292967ad9fd765ff6f
Reviewed-on: https://chromium-review.googlesource.com/191910
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Tested-by: Don Garrett <dgarrett@chromium.org>