cbf: accommodate storm coreboot layout
The cros_bundle_firmware (cbf) tool expects the RO coreboot blob not
to exceed 1MB in size. This has been the case so far, but storm breaks
this convention, at least presently.

Let's take the location of coreboot in the image from the fmap
supplied through the FDT. It remains to be seen if this will work for
all targets. Note that a refactor of this tool if planned for the near
future.

BUG=chrome-os-partner:27784, chromium:384183
TEST=with all other changes landed, invoking

  emerge-storm chromeos-bootimage

  produces a bootable AP148 firmware which starts up, verifies kernel
  image placed on the USB stick, and then loads and starts the kernel.

Change-Id: Ia559e66a532f00df10df4d30fb5e6f9a9983090c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/203548
Reviewed-by: Aaron Durbin <adurbin@chromium.org>