image_to_live.sh: Remove the --no_patch_kernel option.
Whether the kernel is patched or not, it will still have the same vblock.
The --no_patch_kernel option on the devserver is being deprecated and
effectively doesn't do anything, so stop using it.

Also, now that we don't pass --no_patch_kernel to the devserver, we no longer
need our own version of that option. Delete it.

CQ-DEPEND=CL:216578
BUG=chromium:403031
TEST=Ran a butterfly-paladin tryjob with --hwtest.

Change-Id: I8466b882a94dcb1fcf6914504813975391925485
Reviewed-on: https://chromium-review.googlesource.com/216610
Reviewed-by: Yu-Ju Hong <yjhong@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>