gdb_remote: check for symbols availability
When running gdb_remote on a binary, check that the debug symbols are
available for it. If they are not installed, show recommended ways of
installing them then ask whether to continue or not.

This CL also moves all local check before the expensive remote checks
(the host is pingable, etc...).

BUG=chromium:455392
TEST=manual: install a package without debug symbols. run gdb_remote on
one of its binaries.
* gdb_remote warns us and recommend a command.
* Running the said command installs the debug symbols
* Re-running gdb_remote succeeds.

Change-Id: I84f4db92bcd17ce589f2a4d1334f5534af24fcfe
Reviewed-on: https://chromium-review.googlesource.com/247860
Reviewed-by: Bertrand Simonnet <bsimonnet@chromium.org>
Tested-by: Bertrand Simonnet <bsimonnet@chromium.org>
Commit-Queue: Bertrand Simonnet <bsimonnet@chromium.org>