cros_generate_update_payload: Don't force a default chunk_size.
cros_generate_update_payload will always pass a -chunk_size flag to
delta_generator even if none was provided to that script. This patch
changes that to only pass -chunk_size if it was provided by the user.

This makes the effective default value the one set in delta_generator.

BUG=chromium:504447
TEST=Ran cros_generate_update_payload with and without chunk_size.

Change-Id: Ie784e69b8f865a76ef0d00e6b591a4ca630d5000
Reviewed-on: https://chromium-review.googlesource.com/282347
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Queue: Alex Deymo <deymo@chromium.org>