willis: update regexp used to determine branches
The format of the output of 'git branch -vv' has changed with git
version 2.4.0. We need to handle the new format when searching
for branches.

BUG=none
BRANCH=none
TEST=manual

Run willis inside the chroot (git version 2.1.2) and outside
the chroot (2.5.0). Prior to this CL, *every* repo displayed its
master branch with the newer git.

With this CL, both inside and outside show the same information.

Change-Id: I42823f0b38d9cc27897d4ee2e3478bce3725dfa0
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/295685
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>