devserver: Remove old medatada file when a new payload is forced.
./devserver --payload option allows to force a given payload file which
is symlinked to the static directory. Later on, the update.meta file
will be added to the static directory with the metadata of that payload
but will never be removed and regenerated. This patch removes the
update.meta of a forced payload when a new payload is provided.

BUG=None
TEST=`./devserver --payload=foobar.bin` regenerates the metadata

Change-Id: If390785cab1426c9e60e66ac18370240e61eb1da
Reviewed-on: https://chromium-review.googlesource.com/302042
Commit-Ready: Alex Deymo <deymo@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>
Reviewed-by: Simran Basi <sbasi@chromium.org>