bundle_firmware: In RW-CBFS mode, add ecrw and pdrw if available
They won't be in their own fmap regions anymore (with the hash somewhere
else), but both will be in CBFS, with their precalculated hashes in the
CBFS file header.

BUG=chromium:445938
TEST=with all the MULTIPLE_CBFS_INSTANCES stuff enabled, FW_MAIN_[AB]
contain ecrw and pdrw, depending on the hardware in question, both with
a SHA256 hash.

Change-Id: I0f2a4b231ebbae59c729110b0e4e5519df044623
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/306590
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>