cros_generate_update_payload: Fix "cmp" command evaluation linter error.
The construction "if $(cmp ...);" will execute "cmp ..." and then the
output of the cmp command and evaluate the "if" condition based on the
exit code of that output execution. Since cmp doesn't print any output
in either case, the execution of its output doesn't actually execute
anything and the "if" uses the last exit code (from the cmp command
itself).

While this works, it is incorrect. This patch fixes the linter warning
by remoing the $().

BUG=None
TEST=linter.

Change-Id: I1a2c97fb2f3633ea33477acd755087aa5063eb41
Reviewed-on: https://chromium-review.googlesource.com/311107
Reviewed-by: Jason Kusuma <jkusuma@chromium.org>
Tested-by: Alex Deymo <deymo@chromium.org>