bundle_firmware: add a coreboot image with FMAP
As FMAP has become a first class citizen in coreboot allow
one to invoke cbfstool commands on an image that contains
the FMAP data. This is done by creating a coreboot image name
'cb_with_fmap' when the 'coreboot' blob is added instead of
handling it the first time in each pass to _PrepareCbfs().
If the 'cb_with_fmap' is not within the properties when
_PrepareCbfs() is called the processing is deffered. To
achieve this a GetFmap() method was added to the PackFirmware
class.

BUG=chromium:445938
TEST=current cbfstool copy still works
CQ-DEPEND=CL:313781

Change-Id: Ie9be63fe0f621ec914b6e9db32a99441eb801382
Signed-off-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313770
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>